requirements.txt
requests>=2.31.0
pytz>=2024.1
ultralytics>=8.0.0
Pillow>=11.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.14.0
